# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Borsa Batoi ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta usuari autenticat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResource"
                                }
                            }
                        }
                    },
                    "421": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login or password are wrong."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign up",
                "description": "Register user",
                "operationId": "authSignup",
                "requestBody": {
                    "description": "User fields",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta usuari autenticat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dades invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "['email'=>'The field email is required.']"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserStoreRequest": {
                "title": "UserStoreRequest",
                "description": "Registrar Usuari",
                "required": [
                    "email,password,password_confirmation,rol,nombre,domicilio,telefono"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "Email de l'usuari",
                        "type": "email",
                        "example": "pepebotera@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password de l'usuari",
                        "type": "string",
                        "example": "1234"
                    },
                    "password_confirmation": {
                        "title": "password",
                        "description": "Password de l'usuari",
                        "type": "string",
                        "example": "1234"
                    },
                    "rol": {
                        "title": "rol",
                        "description": "Rol de l'usuari",
                        "type": "integer",
                        "example": "2"
                    },
                    "nombre": {
                        "title": "nombre",
                        "description": "Nom de l'alumne o empresa",
                        "type": "string",
                        "example": "Jose Manuel"
                    },
                    "domicilio": {
                        "title": "domicilio",
                        "description": "Adreça de l'alumne",
                        "type": "string",
                        "example": "C/Cid 14"
                    },
                    "telefono": {
                        "title": "telefono",
                        "description": "Telèfon de l'alumne",
                        "type": "string",
                        "example": "666666666"
                    },
                    "cif": {
                        "title": "cif",
                        "description": "Cif de l'empresa",
                        "type": "string",
                        "example": "A12345678"
                    },
                    "localidad": {
                        "title": "localidad",
                        "description": "Localidad de l'empresa",
                        "type": "string",
                        "example": "Alcoi"
                    },
                    "contacto": {
                        "title": "contacto",
                        "description": "Persona de contacte de l'empresa",
                        "type": "string",
                        "example": "Pepe Botera"
                    },
                    "web": {
                        "title": "web",
                        "description": "Pagina web de l'empresa",
                        "type": "string",
                        "example": "http://www.pepebotera.com"
                    },
                    "descripcion": {
                        "title": "descripcion",
                        "description": "Descripció de l'empresa",
                        "type": "string",
                        "example": "Chapuzas a domicilio"
                    },
                    "apellidos": {
                        "title": "apelllidos",
                        "description": "Cognoms de l'alumne",
                        "type": "string",
                        "example": "Miró García"
                    },
                    "info": {
                        "title": "info",
                        "description": "Info de l'alumne",
                        "type": "boolean",
                        "example": "1"
                    },
                    "bolsa": {
                        "title": "bolsa",
                        "description": "bolsa",
                        "type": "boolean",
                        "example": "0"
                    },
                    "cv_enlace": {
                        "title": "cv_enlace",
                        "description": "Enllaç al curriculum de l'alumne",
                        "type": "string",
                        "example": "https://www.pepebotera.com"
                    },
                    "ciclos": {
                        "title": "cicles",
                        "description": "Cicles de l'alumne",
                        "type": "array",
                        "items": {},
                        "example": "[2,4]"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Register"
                }
            },
...
            "LoginResource": {
                "title": "Login Resource",
                "description": "Resposta al Login correcte",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "access_token": {
                                    "title": "access_token",
                                    "description": "Token d'accès",
                                    "type": "string",
                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiI2IiwianRpIjoiNGY4YjE5OTMzNjQxZjVjMDVjNmU1MGE3MGQxNmE5M2ZkYWZmNjkwZDI5M2Y3MTU2NzM5NTkwMDM5YWFhM2FiZDcw"
                                },
                                "rol": {
                                    "title": "rol",
                                    "description": "Rol d'usuari",
                                    "type": "integer",
                                    "example": "2"
                                },
                                "token_type": {
                                    "title": "token_type",
                                    "description": "Tipus de token",
                                    "type": "string",
                                    "example": "Bearer"
                                },
                                "id": {
                                    "title": "id",
                                    "description": "id d'usuari",
                                    "type": "integer",
                                    "example": "12"
                                },
                                "name": {
                                    "title": "name",
                                    "description": "Nom de l'usuari",
                                    "type": "string",
                                    "example": "Ignasi"
                                },
                                "expires_at": {
                                    "title": "expires_at",
                                    "description": "Hora caducitat token",
                                    "type": "time",
                                    "example": "2022-03-17 19:37:34"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LoginResource"
                }
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}